breaks tree tests up into their own files to clean it up in preparation for adding a bunch more
