+++ grunt/tasks/test.js
@@ -3,6 +3,10 @@ module.exports = function test (grunt) {
 	grunt.registerTask('test', 'run jshint, qunit source w/ coverage, and validate HTML',
 		['jshint', 'connect:testServer', 'qunit:noMoment', 'qunit:globals', 'test-dist', 'htmllint']);
 
+	grunt.registerTask('unittest', 'run jshint, qunit source w/ coverage, and validate HTML',
+		['connect:testServer', 'test-dist']);
+
+
 	grunt.registerTask('prep-commonjs-test', 'run commonjs config build and browserify to prep for commonjs test',
 		['commonjs', 'browserify:commonjs']);
